Multi-Function Mic support mod - initial commit
MFmic is a protocol system that allows an HM-151 or equivalent mic to send extended commands to the RDU.  It shares the CLI UART port.  The protocol uses a wired-remote configured HM-133 interface and an ATTINY1604 to move mic button data to the RDU via the up/dn signal of the IC-900 mic connector and reproduce the up/dn signals if the com port is in use.  Also seamlessly supports a standard HM-14 mic.